requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
confusable-homoglyphs==3.2.0
Django==3.2.7
django-bootstrap4==3.0.1
django-registration==3.2
django-registration-redux==2.9
django-tinymce4-lite==1.8.0
djangorestframework==3.12.4
jsmin==3.0.0
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.2
